requirements.txt
pymupdf
torch
transformers
scikit-learn
numpy
pillow
langchain
langchain-core
langchain-community
langchain-ollama
faiss-cpu
python-dotenv